# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Book Management API",
        "description": "Comprehensive API documentation untuk sistem manajemen buku dengan Laravel dan Sanctum authentication. Menyediakan fitur CRUD untuk books, user management, dan sistem peminjaman buku dengan validasi stock.",
        "contact": {
            "email": "agung.alfatah43@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/loans": {
            "get": {
                "tags": [
                    "Book Loans"
                ],
                "summary": "Get all loans (admin only)",
                "description": "Daftar semua peminjaman buku (admin only)",
                "operationId": "baf3364ec5b1a481a6bc6e3791ae38b7",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by loan status",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "borrowed",
                                "returned",
                                "overdue",
                                "all"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Loans retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Loans retrieved successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Admin access required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Admin access required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Book Loans"
                ],
                "summary": "Borrow a book",
                "description": "User meminjam buku dengan validasi stock",
                "operationId": "89b59f574651a4c4935318feca9455eb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BorrowBookRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book borrowed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Book borrowed successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/BookLoan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Book not available or already borrowed by user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Book is not available for loan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/loans/my-loans": {
            "get": {
                "tags": [
                    "Book Loans"
                ],
                "summary": "Get current user's loans",
                "description": "Get all loans for the authenticated user",
                "operationId": "87f69a3fcd7a3c53632bb2537535a2dc",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by loan status (borrowed, returned, overdue, all)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "borrowed",
                            "enum": [
                                "borrowed",
                                "returned",
                                "overdue",
                                "all"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User loans retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Your loans retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/BookLoan"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/loans/{user_id}": {
            "get": {
                "tags": [
                    "Book Loans"
                ],
                "summary": "Get user's borrowed books by user ID",
                "description": "Daftar buku yang sedang dipinjam oleh user",
                "operationId": "18af5a12170954111478e6d53be058cc",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by loan status (borrowed, returned, or all)",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "borrowed",
                                "returned",
                                "all"
                            ],
                            "example": "borrowed"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User loans retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User loans retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/BookLoan"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/loans/{loan_id}/return": {
            "put": {
                "tags": [
                    "Book Loans"
                ],
                "summary": "Return a borrowed book",
                "description": "User mengembalikan buku yang dipinjam",
                "operationId": "a8801ad8d98aed8277d85701b9a94e66",
                "parameters": [
                    {
                        "name": "loan_id",
                        "in": "path",
                        "description": "Loan ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "550e8400-e29b-41d4-a716-446655440000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book returned successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Book returned successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/BookLoan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Book already returned or invalid loan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Book has already been returned"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BookLoan": {
                "title": "Book Loan",
                "description": "Book loan model",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "user_id": {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "book_id": {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "loan_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2023-01-01"
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2023-01-15"
                    },
                    "return_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2023-01-10",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "borrowed",
                            "returned"
                        ],
                        "example": "borrowed"
                    },
                    "is_overdue": {
                        "type": "boolean",
                        "example": false
                    },
                    "days_until_due": {
                        "type": "integer",
                        "example": 5
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "book": {
                        "$ref": "#/components/schemas/Book"
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2023-01-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2023-01-01T00:00:00.000000Z"
                    }
                },
                "type": "object"
            },
...
            "BorrowBookRequest": {
                "title": "Borrow Book Request",
                "description": "Request payload untuk meminjam buku",
                "required": [
                    "book_id"
                ],
                "properties": {
                    "book_id": {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "loan_days": {
                        "description": "Number of days to borrow (default: 14)",
                        "type": "integer",
                        "example": 14
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
...
            "Unauthenticated": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/UnauthenticatedResponse"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}